Address feedback. Split off paging types into core-paging
