+++ sdk/core/core-paging/ThirdPartyNotices.txt
@@ -1,0 +1,35 @@
+Third Party Notices for ms-rest-js
+
+This project incorporates material from the project(s) listed below (collectively, Third Party Code).
+Microsoft, Inc. Microsoft is not the original author of the Third Party Code.
+The original copyright notice and license, under which Microsoft received such Third Party Code,
+are set out below.  This Third Party Code is licensed to you under their original license terms set forth below.
+Microsoft reserves all other rights not expressly granted, whether by implication, estoppel or otherwise.
+
+1. uuid (https://github.com/kelektiv/node-uuid)
+
+%% uuid NOTICES AND INFORMATION BEGIN HERE
+=========================================
+The MIT License (MIT)
+
+Copyright (c) 2010-2016 Robert Kieffer and other contributors
+
+Permission is hereby granted, free of charge, to any person obtaining a copy
+of this software and associated documentation files (the "Software"), to deal
+in the Software without restriction, including without limitation the rights
+to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+copies of the Software, and to permit persons to whom the Software is
+furnished to do so, subject to the following conditions:
+
+The above copyright notice and this permission notice shall be included in all
+copies or substantial portions of the Software.
+
+THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+SOFTWARE.
+=========================================
+END OF uuid NOTICES AND INFORMATION
