+++ sdk/keyvault/keyvault-secrets/package.json
@@ -26,10 +26,12 @@
   "dependencies": {
     "@azure/ms-rest-azure-js": "^1.3.2",
     "@azure/ms-rest-js": "^1.2.6",
+    "@azure/core-paging": "^1.0.0",
     "tslib": "^1.9.3"
   },
   "devDependencies": {
     "@azure/ms-rest-nodeauth": "^0.9.2",
+    "@azure/core-paging": "^1.0.0",
     "@types/chai": "^4.1.6",
     "@types/mocha": "^5.2.5",
     "chai": "^4.2.0",
